+++ src/main/kotlin/com/github/siropkin/kursor/TextIndicatorVerticalPositions.kt
@@ -1,7 +1,0 @@
-package com.github.siropkin.kursor
-
-object TextIndicatorVerticalPositions {
-    const val TOP = "top"
-    const val MIDDLE = "middle"
-    const val BOTTOM = "bottom"
-}